flask/requirements.txt
alembic==1.4.2
click==7.1.2
dnspython==1.16.0
eventlet==0.25.2
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-SocketIO==4.3.0
Flask-SQLAlchemy==2.4.1
greenlet==0.4.15
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.2
MarkupSafe==1.1.1
monotonic==1.5
psycopg2-binary==2.8.5
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
python-engineio==3.12.1
python-socketio==4.5.1
six==1.14.0
SQLAlchemy==1.3.16
Werkzeug==1.0.1
